+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # Releases
+## 0.0.3 - Name Uniqueness Validation
+- Now the program validates the names of all structures in the arrays of the provided file. All of the names must be unique.
+
 ## 0.0.2 - Bugfixes
 - Fixes the validation of items in the `"outputBranches"` array. It must contain elements.
 - Fixes the error displayed when validating  the `"type"` of an `"inputParameters"` element. It used to report an `undefined` error, but now it reports a proper message.